% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LerArqXlsx.R
\name{GerarTable}
\alias{GerarTable}
\title{Generate a TABLE comparing the XLSX files read}
\usage{
GerarTable(Tabela1, Tabela2)
}
\arguments{
\item{Tabela1}{One of the tables read with the LerTable function}

\item{Tabela2}{The other table read with the LerTable function

#' @examples
GerarEx <- GerarTable(T1, T2)

GerarEx}
}
\description{
Generate a TABLE comparing the XLSX files read
}
