% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LerArqXlsx.R
\name{LerTable}
\alias{LerTable}
\title{Read XLSX tables related to the OPPA project}
\usage{
LerTable(path, TabelaBase = 0)
}
\arguments{
\item{path}{The file's path to the XSLX file.}

\item{TabelaBase}{The source table that will be compared in GerarTable}
}
\description{
Read XLSX tables related to the OPPA project
}
\examples{
lerTabNormalEx <- ReadExcell(system.file("extdata/examples/ExSource.xls",
                                                package = "OPPA"))
lerTabNormalEx

lerTabBaseEx <- ReadExcell(system.file("extdata/examples/ExSource.xls",
                                                package = "OPPA"))
lerTabBaseEx

}
