% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LerArqXlsx.R
\name{OPPA}
\alias{OPPA}
\title{Read XLSX files related to the OPPA project}
\usage{
OPPA(path)
}
\arguments{
\item{path}{The file's path to the XSLX file.}
}
\value{
A data frame.
}
\description{
Read XLSX files related to the OPPA project
}
\details{
By now, this function suports any XLSX file.
}
\examples{
# these datasets are real ones. But it's here just for test purpose.

# example selecting the file
OPPA("E:/Back-up/Projetos - FNO/R projects/Ic/OPPA/inst/extdata/arquivo.xlsx")

class(OPPA)
}
