% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oppa.R
\name{read_oppa}
\alias{read_oppa}
\title{Read XLSX files related to the OPPA project}
\usage{
read_oppa(path)
}
\arguments{
\item{path}{The file's path to the XSLX file.}
}
\value{
A data frame.
}
\description{
Read XLSX files related to the OPPA project
}
\details{
By now, this function suports any XLSX file.
}
\examples{
# these datasets aren't real ones. But it's here just for test purpose.

oppa <- read_oppa(system.file("extdata/arquivos/oppa.xlsx",
                              package = "OPPA"))
}
