% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LerArqXlsx.R
\name{ReadExcell}
\alias{ReadExcell}
\title{Read XLSX files related to the OPPA project}
\usage{
ReadExcell(path)
}
\arguments{
\item{path}{The file's path to the XSLX file.}
}
\value{
A data frame.
}
\description{
Read XLSX files related to the OPPA project
}
\details{
By now, this function suports any XLSX file.
}
\examples{
lerEx <- ReadExcell(system.file("extdata/examples/ExSource.xls",
                                                package = "OPPA"))
lerEx
}
